Move ixmp.reporting.utils tests to a new file
